LLM_SEND_EMAIL/requirements.txt
ffmpeg-python==0.2.0
langchain==0.0.351
openai-whisper==20231117
streamlit==1.29.0
openai==1.5.0
langchain_community